working on assessment router
