+++ server/routes/cohortRoutes.js
@@ -1,5 +1,5 @@
 const express = require('express')
-const db = require('../db')
+const db = require('../db/cohorts')
 const router = express.Router()
 
 module.exports = router
