+++ server/server.js
@@ -5,6 +5,7 @@ const server = express()
 
 const starterRoute = require('./routes/starter')
 const cohortRoutes = require('./routes/cohortRoutes')
+const assessmentsRoute = require('./routes/assessmentsRoute')
 
 server.use(express.json())
 server.use(express.static(path.join(__dirname, './public')))
